Replace grid list with map of cells
Using a list to represent the grid of cells is inefficient for
large mazes. Random access is not available to lists. Instead,
the cells are now represented in a map, where the key is the
`{row, col}` tuple and the value is the type of cell. This
allows us to access individual elements quickly given their
coordinates.